simplify exception handler
