--- lib/march_hare/exception_handler/forgiving_exception_handler.rb
@@ -1,5 +1,0 @@
-module MarchHare
-  class ForgivingExceptionHandler < com.rabbitmq.client.impl.ForgivingExceptionHandler
-    include RubifyJavaExceptionHandler
-  end
-end
