--- lib/march_hare/exception_handler/strict_exception_handler.rb
@@ -1,5 +1,0 @@
-module MarchHare
-  class StrictExceptionHandler < com.rabbitmq.client.impl.StrictExceptionHandler
-    include RubifyJavaExceptionHandler
-  end
-end
